requirements.txt
tensorflow==2.18.0
matplotlib==3.8.2
numpy==1.26.4
jupyter==1.0.0
pillow==11.1.0
scikit-learn==1.5.2